Repair context types, add Role context, add CrossGuild annotation (#137)
In a previous PR, some contexts were not changed from **IssuerOnly** to **IssuerAware**. This has been patched.

I've added a context to resolve a Role. It makes use of the next item I added.

The CrossGuild annotation is used to signify whether or not the context is to be interpreted from a guild perspective or a bot perspective (i.e. all guilds). This annotation could probably be better named.