Do not filter out valid completion arguments
Fixes a bug introduced by PR 200. The bug related to the Values annotation which uses this getCompletionValues method also.

This is definitely not the propersolution. The underlying issue is
related to the logic calling `args` which is the original argument
array. And it should not be using arguments which have been resolved in
a previous parameter.
I was not able to find a quick/simple way to accommodate that issue.