Deprecate bukkit command manager getTimings and remove creation of timing calls #385
There is an acknowledged loss in functionality with this approach for older versions of
Minecraft and also Bukkit/Spigot servers which still have continued access to timings.
This change was implemented on the current acf snapshot version knowing this,
to enable plugin authors to simply recompile to fix this issue without making
the small change of upgrading the dependency.